pyproject.toml
[tool.poetry]
name = "allyoucanuse"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.0"
jsonlines = "^3.0.0"
pendulum = "^2.1.2"
PyYAML = "^6.0"
python-benedict = "^0.24.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
